+++ uguid/CHANGELOG.md
@@ -1,8 +1,9 @@
-# 2.1.1
+# Unreleased
 
-* Fix non-upper-case-globals linter warning.
+* Added `Variant` enum and `Guid::variant` method.
 * Conversions of the `time_low` field to/from bytes now treat that field
   as native endian rather than little endian.
+* Fix non-upper-case-globals linter warning.
 
 # 2.1.0