+++ uguid/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 * Added `Variant` enum and `Guid::variant` method.
+* Added `Guid::version` method.
 * Conversions of the `time_low` field to/from bytes now treat that field
   as native endian rather than little endian.
 * Fix non-upper-case-globals linter warning.